Move Timer to shared CoreLib partition.
